+++ packages/store-engine/src/main/engine/IndexedDBEngine.test.browser.js
@@ -58,13 +58,13 @@ describe('IndexedDBEngine', () => {
   });
 
   describe('"append"', () => {
-    Object.entries(require('../../test/shared/append')).map(([description, testFunction]) => {
+    Object.entries(require('../test/shared/append')).map(([description, testFunction]) => {
       it(description, done => testFunction(done, engine));
     });
   });
 
   describe('"create"', () => {
-    Object.entries(require('../../test/shared/create')).map(([description, testFunction]) => {
+    Object.entries(require('../test/shared/create')).map(([description, testFunction]) => {
       it(description, done => testFunction(done, engine));
     });
 
@@ -100,13 +100,13 @@ describe('IndexedDBEngine', () => {
   });
 
   describe('"delete"', () => {
-    Object.entries(require('../../test/shared/delete')).map(([description, testFunction]) => {
+    Object.entries(require('../test/shared/delete')).map(([description, testFunction]) => {
       it(description, done => testFunction(done, engine));
     });
   });
 
   describe('"deleteAll"', () => {
-    Object.entries(require('../../test/shared/deleteAll')).map(([description, testFunction]) => {
+    Object.entries(require('../test/shared/deleteAll')).map(([description, testFunction]) => {
       it(description, done => testFunction(done, engine));
     });
   });
@@ -159,37 +159,37 @@ describe('IndexedDBEngine', () => {
   });
 
   describe('"purge"', () => {
-    Object.entries(require('../../test/shared/purge')).map(([description, testFunction]) => {
+    Object.entries(require('../test/shared/purge')).map(([description, testFunction]) => {
       it(description, done => testFunction(done, engine, initEngine));
     });
   });
 
   describe('"read"', () => {
-    Object.entries(require('../../test/shared/read')).map(([description, testFunction]) => {
+    Object.entries(require('../test/shared/read')).map(([description, testFunction]) => {
       it(description, done => testFunction(done, engine));
     });
   });
 
   describe('"readAll"', () => {
-    Object.entries(require('../../test/shared/readAll')).map(([description, testFunction]) => {
+    Object.entries(require('../test/shared/readAll')).map(([description, testFunction]) => {
       it(description, done => testFunction(done, engine));
     });
   });
 
   describe('"readAllPrimaryKeys"', () => {
-    Object.entries(require('../../test/shared/readAllPrimaryKeys')).map(([description, testFunction]) => {
+    Object.entries(require('../test/shared/readAllPrimaryKeys')).map(([description, testFunction]) => {
       it(description, done => testFunction(done, engine));
     });
   });
 
   describe('"update"', () => {
-    Object.entries(require('../../test/shared/update')).map(([description, testFunction]) => {
+    Object.entries(require('../test/shared/update')).map(([description, testFunction]) => {
       it(description, done => testFunction(done, engine));
     });
   });
 
   describe('"updateOrCreate"', () => {
-    Object.entries(require('../../test/shared/updateOrCreate')).map(([description, testFunction]) => {
+    Object.entries(require('../test/shared/updateOrCreate')).map(([description, testFunction]) => {
       it(description, done => testFunction(done, engine));
     });
   });
