+++ packages/store-engine/src/main/test/append.ts
@@ -1,4 +1,4 @@
-import {CRUDEngine} from '../main/engine';
+import {CRUDEngine} from '../engine';
 
 const TABLE_NAME = 'the-simpsons';
