+++ packages/store-engine/src/main/test/readAllPrimaryKeys.ts
@@ -17,12 +17,12 @@
  *
  */
 
-import {CRUDEngine} from '../main/engine';
+import {CRUDEngine} from '../engine';
 
 const TABLE_NAME = 'the-simpsons';
 
 export default {
-  'deletes all records from a database table.': (done: DoneFn, engine: CRUDEngine) => {
+  'gets the primary keys of all records in a table.': (done: DoneFn, engine: CRUDEngine) => {
     const homer = {
       entity: {
         firstName: 'Homer',
@@ -47,18 +47,19 @@ export default {
       primaryKey: 'marge-simpson',
     };
 
+    const allEntities = [homer, lisa, marge];
+
     Promise.all([
       engine.create(TABLE_NAME, homer.primaryKey, homer.entity),
       engine.create(TABLE_NAME, lisa.primaryKey, lisa.entity),
       engine.create(TABLE_NAME, marge.primaryKey, marge.entity),
     ])
-      .then(() => engine.deleteAll(TABLE_NAME))
-      .then(hasBeenDeleted => {
-        expect(hasBeenDeleted).toBe(true);
-        return engine.readAllPrimaryKeys(TABLE_NAME);
-      })
+      .then(() => engine.readAllPrimaryKeys(TABLE_NAME))
       .then(primaryKeys => {
-        expect(primaryKeys.length).toBe(0);
+        expect(primaryKeys.length).toBe(allEntities.length);
+        for (const counter in allEntities) {
+          expect(primaryKeys[counter]).toBe(allEntities[counter].primaryKey);
+        }
         done();
       })
       .catch(error => done.fail(error));
