+++ packages/store-engine/src/main/test/shared/deleteAll.js
