+++ packages/store-engine/src/main/test/updateOrCreate.ts
@@ -17,41 +17,53 @@
  *
  */
 
-import {CRUDEngine} from '../main/engine';
-import RecordNotFoundError from '../main/engine/error/RecordNotFoundError';
+import {CRUDEngine} from '../engine';
 
 const TABLE_NAME = 'the-simpsons';
 
 interface DomainEntity {
-  some: string;
+  name: string;
 }
 
 export default {
-  'returns a database record.': (done: DoneFn, engine: CRUDEngine) => {
+  'creates a record if it does not exist in the database.': (done: DoneFn, engine: CRUDEngine) => {
     const PRIMARY_KEY = 'primary-key';
 
     const entity = {
-      some: 'value',
+      name: 'Old monitor',
+    };
+
+    const expectedAmountOfProperties = 1;
+
+    engine
+      .updateOrCreate(TABLE_NAME, PRIMARY_KEY, entity)
+      .then(primaryKey => engine.read<DomainEntity>(TABLE_NAME, primaryKey))
+      .then(updatedRecord => {
+        expect(updatedRecord.name).toBe(entity.name);
+        expect(Object.keys(updatedRecord).length).toBe(expectedAmountOfProperties);
+        done();
+      })
+      .catch(done.fail);
+  },
+  'updates an existing database record.': (done: DoneFn, engine: CRUDEngine) => {
+    const PRIMARY_KEY = 'primary-key';
+
+    const entity = {
+      name: 'Old monitor',
+    };
+
+    const update = {
+      name: 'Old monitor2',
     };
 
     engine
       .create(TABLE_NAME, PRIMARY_KEY, entity)
+      .then(() => engine.updateOrCreate(TABLE_NAME, PRIMARY_KEY, update))
       .then(primaryKey => engine.read<DomainEntity>(TABLE_NAME, primaryKey))
-      .then(record => {
-        expect(record.some).toBe(entity.some);
+      .then(updatedRecord => {
+        expect(updatedRecord.name).toBe(update.name);
         done();
       })
-      .catch(error => done.fail(error));
-  },
-  'throws an error if a record cannot be found.': (done: DoneFn, engine: CRUDEngine) => {
-    const PRIMARY_KEY = 'primary-key';
-
-    engine
-      .read(TABLE_NAME, PRIMARY_KEY)
-      .then(() => done.fail(new Error('Method is supposed to throw an error.')))
-      .catch(error => {
-        expect(error).toEqual(jasmine.any(RecordNotFoundError));
-        done();
-      });
+      .catch(done.fail);
   },
 };
