+++ packages/store-engine/tsconfig.json
@@ -1,6 +1,5 @@
 {
   "compilerOptions": {
-    "allowJs": true,
     "outDir": "dist/commonjs",
     "rootDir": "src/main"
   },
